pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ds-resource-plugin-py-lib"
version = "0.1.0-alpha.2"
description = "A Python package from the ds-common library collection"
dependencies = [
    "pandas>=2.1.0,<3.0.0",
    "awswrangler>=3.11.0,<4.0.0",
    "pyyaml>=6.0.2,<7.0.0",
    "typing-extensions>=4.0.0,<5.0.0",
    "ds-common-logger-py-lib>=0.1.0a1,<1.0.0",
    "ds-common-serde-py-lib>=0.1.0a2,<1.0.0",
]
requires-python = ">=3.11"
license = "Apache-2.0"
license-files = ["LICENSE-APACHE"]
keywords = ["ds", "common", "python"]
authors = [
    {name = "Kristoffer Varslott", email = "hello@grasplabs.com"}
]
maintainers = [
    {name = "Kristoffer Varslott", email = "hello@grasplabs.com"}
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Operating System :: OS Independent",
  "Intended Audience :: Developers",
  "Topic :: Software Development :: Libraries :: Python Modules",
]
readme = "PyPI.md"

[project.urls]
Documentation = "https://grasp-labs.github.io/ds-resource-plugin-py-lib/"
Repository = "https://github.com/grasp-labs/ds-resource-plugin-py-lib/"
Issues = "https://github.com/grasp-labs/ds-resource-plugin-py-lib/issues/"
Changelog = "https://github.com/grasp-labs/ds-resource-plugin-py-lib/releases"

[project.optional-dependencies]
dev = [
    "ruff>=0.1.8",
    "mypy>=1.7.0",
    "pandas-stubs>=2.1.0",
    "types-PyYAML>=6.0.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.3.0",
    "pre-commit>=3.6.0",
    "build>=1.0.0",
    "twine>=4.0.0",
    "sphinx>=7.1.0",
    "sphinx-autoapi>=3.0.0",
    "sphinx-material>=0.0.35",
]
test = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.3.0",
]
docs = [
    "sphinx>=7.1.0",
    "sphinx-autoapi>=3.0.0",
    "sphinx-material>=0.0.35",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
ds_resource_plugin_py_lib = ["py.typed"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"
